# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Vakt API
  version: "0.23.0"
  description: |
    Self-hosted Security & Compliance Platform API.
    All data stays within your infrastructure — no phone-home, no cloud dependency.

    **Authentication:** Paseto v2 bearer tokens. Obtain a token via `POST /auth/login`.
    Include it as `Authorization: Bearer <token>` on all protected endpoints.

    **Base URL:** `/api/v1`

    **Pagination:** `?page=1&limit=25` on list endpoints.

    **Error format:**
    ```json
    { "error": "message", "code": "ERROR_CODE", "details": {} }
    ```
servers:
  - url: /api/v1
    description: Vakt API v1
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Paseto
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 25
        maximum: 100
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    Asset:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [web_app, server, database, container, repo]
        target:
          type: string
        criticality:
          type: string
          enum: [low, medium, high, critical]
        tags:
          type: array
          items:
            type: string
        environment:
          type: string
          enum: [prod, staging, dev]
        created_at:
          type: string
          format: date-time
    CreateAssetRequest:
      type: object
      required: [name, type, target, criticality]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [web_app, server, database, container, repo]
        target:
          type: string
        criticality:
          type: string
          enum: [low, medium, high, critical]
        tags:
          type: array
          items:
            type: string
    Finding:
      type: object
      properties:
        id:
          type: string
          format: uuid
        asset_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [info, low, medium, high, critical]
        status:
          type: string
          enum: [open, in_progress, resolved, accepted]
        cve:
          type: string
        cvss_score:
          type: number
        created_at:
          type: string
          format: date-time

    # ── SecVitals ─────────────────────────────────────────────────────────────
...
    Risk:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        likelihood:
          type: integer
          minimum: 1
          maximum: 5
        impact:
          type: integer
          minimum: 1
          maximum: 5
        risk_score:
          type: integer
        status:
          type: string
          enum: [open, mitigated, accepted, closed]
        owner:
          type: string
    Incident:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [open, investigating, contained, resolved]
        type:
          type: string
        created_at:
          type: string
          format: date-time

    # ── SecPrivacy ────────────────────────────────────────────────────────────
...
paths:
...
  /vaktscan/assets:
    get:
      summary: List assets
      tags: [vaktscan]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated asset list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Asset'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      summary: Create an asset
      tags: [vaktscan]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetRequest'
      responses:
        '201':
          description: Asset created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '400':
          description: Validation error
...
  /vaktscan/findings:
    get:
      summary: List findings
      tags: [vaktscan]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: asset_id
          in: query
          schema:
            type: string
            format: uuid
        - name: severity
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated findings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Finding'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
  /vaktcomply/risks:
    get:
      summary: List risks
      tags: [vaktcomply]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Risk list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Risk'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      summary: Create a risk
      tags: [vaktcomply]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Risk'
      responses:
        '201':
          description: Risk created
...
  /vaktcomply/incidents:
    get:
      summary: List incidents
      tags: [vaktcomply]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Incident list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Incident'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      summary: Create an incident
      tags: [vaktcomply]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Incident'
      responses:
        '201':
          description: Incident created
...
  /vaktcomply/policies:
    get:
      summary: List policies
      tags: [vaktcomply]
      responses:
        '200':
          description: List of policies
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string, format: uuid }
                    title: { type: string }
                    status: { type: string, enum: [draft, active, archived] }
                    version: { type: integer }
                    owner: { type: string }
                    review_date: { type: string, format: date }
                    created_at: { type: string, format: date-time }
    post:
      summary: Create policy
      tags: [vaktcomply]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, content]
              properties:
                title: { type: string }
                content: { type: string }
                owner: { type: string }
                review_interval_days: { type: integer }
      responses:
        '201':
          description: Created policy
...
  /vaktcomply/capas:
    get:
      summary: List corrective/preventive actions (CAPAs)
      tags: [vaktcomply]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [open, in_progress, closed, verified] }
      responses:
        '200':
          description: List of CAPAs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string, format: uuid }
                    title: { type: string }
                    status: { type: string }
                    priority: { type: string, enum: [low, medium, high, critical] }
                    due_date: { type: string, format: date }
                    assignee: { type: string }
                    source_type: { type: string }
    post:
      summary: Create CAPA
      tags: [vaktcomply]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, priority]
              properties:
                title: { type: string }
                description: { type: string }
                priority: { type: string, enum: [low, medium, high, critical] }
                due_date: { type: string, format: date }
                assignee_id: { type: string, format: uuid }
                source_type: { type: string }
                source_id: { type: string }
      responses:
        '201':
          description: Created CAPA
...
  /vaktprivacy/dsr:
    get:
      summary: List data subject requests (Art. 15–22 DSGVO)
      tags: [vaktprivacy]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [open, in_progress, completed, rejected] }
        - name: type
          in: query
          schema: { type: string, enum: [access, erasure, rectification, portability, restriction, objection] }
      responses:
        '200':
          description: List of DSRs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id: { type: string, format: uuid }
                    type: { type: string }
                    status: { type: string }
                    subject_name: { type: string }
                    subject_email: { type: string }
                    received_at: { type: string, format: date-time }
                    deadline: { type: string, format: date }
    post:
      summary: Create DSR record
      tags: [vaktprivacy]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, subject_email]
              properties:
                type: { type: string, enum: [access, erasure, rectification, portability, restriction, objection] }
                subject_name: { type: string }
                subject_email: { type: string, format: email }
                description: { type: string }
      responses:
        '201':
          description: Created DSR
...